closes #42 closes #43, gui ready for all new configData and Actions

+++ src/main/java/edu/mscd/thesis/controller/GameConfig.java
@@ -4,7 +4,7 @@ public interface GameConfig {
 	
 	public boolean isPaused();
 	public boolean isStep();
-	public int getSpeed();
+	public double getSpeed();
 	public AiMode getAiMode();
 	
 	public GameConfig copy();
